Dashboard/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
paho-mqtt>=1.6.1
influxdb-client>=1.18.0
requests>=2.31.0
websockets>=11.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0